pyproject.toml
[tool.poetry]
package-mode = false
name = "hri"
version = "0.1.0"
description = ""
authors = ["Dan <dan@loopsites.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
tabulate = "^0.9.0"
matplotlib = "^3.9.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
flake8-bugbear = "^24.8.19"
isort = "^5.13.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
